requirements.txt
pillow
tk_html_widgets
